[refactor] Preview: move to the androidx package for Android Studio to work
