+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/components/AppBottomNavigationBar.kt
@@ -9,6 +9,7 @@
 
 package com.rwmobi.kunigami.ui.components
 
+import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
@@ -31,7 +32,6 @@ import androidx.navigation.compose.rememberNavController
 import com.rwmobi.kunigami.ui.navigation.AppNavigationItem
 import com.rwmobi.kunigami.ui.theme.AppTheme
 import com.rwmobi.kunigami.ui.theme.getDimension
-import com.rwmobi.kunigami.ui.utils.Preview
 import kunigami.composeapp.generated.resources.Res
 import kunigami.composeapp.generated.resources.content_description_navigation_bar
 import org.jetbrains.compose.resources.ExperimentalResourceApi
