+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/components/AppNavigationRail.kt
@@ -9,6 +9,7 @@
 
 package com.rwmobi.kunigami.ui.components
 
+import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -32,7 +33,6 @@ import androidx.navigation.compose.rememberNavController
 import com.rwmobi.kunigami.ui.navigation.AppNavigationItem
 import com.rwmobi.kunigami.ui.theme.AppTheme
 import com.rwmobi.kunigami.ui.theme.getDimension
-import com.rwmobi.kunigami.ui.utils.Preview
 import kunigami.composeapp.generated.resources.Res
 import kunigami.composeapp.generated.resources.content_description_navigation_rail
 import org.jetbrains.compose.resources.ExperimentalResourceApi
