+++ composeApp/src/desktopMain/kotlin/com/rwmobi/kunigami/ui/components/LazyColumnMultiplatform.desktop.kt
@@ -26,9 +26,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
+import com.rwmobi.kunigami.ui.previewparameter.Preview
 import com.rwmobi.kunigami.ui.theme.AppTheme
 import com.rwmobi.kunigami.ui.theme.getDimension
-import com.rwmobi.kunigami.ui.utils.Preview
 
 @Composable
 actual fun ScrollbarMultiplatform(
